i add some validations

+++ src/controller/ingredients.ts
@@ -2,7 +2,12 @@ import { serverError } from "../Middleware/serverError";
 import { NextFunction, Request, Response } from "express";
 import Ingredient from "../models/Ingredient";
 import Recipe from "../models/Recipe";
-export const getAllingredients = async (
+import mongoose from "mongoose";
+import { invaldCredentialsErrorHandler } from "../Middleware/errors";
+
+const expectedIngredientFields = ["name", "recipes"];
+
+export const getAllIngredients = async (
   req: Request,
   res: Response,
   next: NextFunction
@@ -11,16 +16,45 @@ export const getAllingredients = async (
     const ingredients = await Ingredient.find().populate("recipes");
     res.status(200).json(ingredients);
   } catch (error) {
-    console.log("the error from getAllingredients", error);
+    console.log("the error from getAllIngredients", error);
     return next(serverError);
   }
 };
+
 export const createIngredient = async (
   req: Request,
   res: Response,
   next: NextFunction
 ) => {
   try {
+    const bodyKeys = Object.keys(req.body || {});
+    for (const key of bodyKeys) {
+      if (!expectedIngredientFields.includes(key)) {
+        return next(
+          invaldCredentialsErrorHandler(
+            `Invalid field "${key}" in request body. Did you mean one of: ${expectedIngredientFields.join(
+              ", "
+            )}?`
+          )
+        );
+      }
+    }
+
+    if (!req.body.name) {
+      return next(invaldCredentialsErrorHandler("Ingredient name is required"));
+    }
+
+    const existingIngredient = await Ingredient.findOne({
+      name: req.body.name,
+    });
+    if (existingIngredient) {
+      return next(
+        invaldCredentialsErrorHandler(
+          "Ingredient with this name already exists"
+        )
+      );
+    }
+
     const ingredient = await Ingredient.create(req.body);
     const save = await ingredient.save();
     if (save.recipes.length) {
@@ -35,12 +69,17 @@ export const createIngredient = async (
     next(serverError);
   }
 };
+
 export const deleteIngredient = async (
   req: Request,
   res: Response,
   next: NextFunction
 ) => {
   try {
+    if (!mongoose.Types.ObjectId.isValid(req.params.id!)) {
+      return next(invaldCredentialsErrorHandler("Invalid ingredient ID"));
+    }
+
     const deleted = await Ingredient.findByIdAndDelete(req.params.id);
     if (!deleted)
       return res.status(404).json({ message: "ingredient not found" });
